refactor backend API route code
